Fix: merge sorald test with githubmode tests
